fix(cypress): add Chrome launch flags to prevent renderer crashes in CI
GitHub Actions runners default to 64MB /dev/shm, which Chrome headless
uses for shared memory. Chrome 150 exceeds this limit during test runs,
causing renderer process crashes across multiple test files.

This adds a before:browser:launch hook with:
- --disable-dev-shm-usage: writes to /tmp instead of /dev/shm
- --disable-gpu: reduces memory overhead in headless mode
- --js-flags=--max-old-space-size=4096: increases V8 heap limit

Reverts the previous file-splitting approach since the root cause was
Chrome's shared memory, not test file size.
