+++ packages/cypress/cypress.config.ts
@@ -93,6 +93,15 @@ export default defineConfig({
       coverage(on, config);
       setupWebsockets(on, config);
 
+      on('before:browser:launch', (browser, launchOptions) => {
+        if (browser.family === 'chromium') {
+          launchOptions.args.push('--disable-dev-shm-usage');
+          launchOptions.args.push('--disable-gpu');
+          launchOptions.args.push('--js-flags=--max-old-space-size=4096');
+        }
+        return launchOptions;
+      });
+
       on('task', {
         readJSON(filePath: string) {
           const absPath = path.resolve(__dirname, filePath);
@@ -148,15 +157,12 @@ export default defineConfig({
       // Delete videos for specs without failing or retried tests
       on('after:spec', (_, results) => {
         if (results.video) {
-          // Keep the video when results.tests is null (Chrome renderer crash)
-          const failures =
-            // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-            !results.tests ||
-            results.tests.some((test) =>
-              test.attempts[config.env.MOCK ? 'every' : 'some'](
-                (attempt) => attempt.state === 'failed',
-              ),
-            );
+          // Do we have failures for any retry attempts?
+          const failures = results.tests.some((test) =>
+            test.attempts[config.env.MOCK ? 'every' : 'some'](
+              (attempt) => attempt.state === 'failed',
+            ),
+          );
           if (!failures) {
             // delete the video if the spec passed and no tests retried
             fs.unlinkSync(results.video);
@@ -170,19 +176,13 @@ export default defineConfig({
       });
 
       on('after:run', async () => {
-        try {
-          await afterRunHook();
-        } catch (e) {
-          console.warn('mochawesome report generation failed:', e);
-        }
+        // cypress-mochawesome-reporter
+        await afterRunHook();
 
-        try {
-          const outputFile = path.join(__dirname, resultsDir, 'junit-report.xml');
-          const inputFiles = [`./${resultsDir}/junit/*.xml`];
-          await mergeFiles(outputFile, inputFiles);
-        } catch (e) {
-          console.warn('junit report merge failed:', e);
-        }
+        // merge junit reports into a single report
+        const outputFile = path.join(__dirname, resultsDir, 'junit-report.xml');
+        const inputFiles = [`./${resultsDir}/junit/*.xml`];
+        await mergeFiles(outputFile, inputFiles);
       });
 
       // Apply retries only for tests in the "e2e" folder. 2 retries by default, after a test failure.
