+++ packages/cypress/cypress/tests/mocked/gpuaas/infrastructure.cy.ts
@@ -383,4 +383,265 @@ describe('GPUaaS Infrastructure Page', () => {
       infrastructurePage.findBorrowingLendingEmptyState().should('exist');
     });
   });
+
+  describe('Cluster queue utilization section', () => {
+    beforeEach(() => {
+      asClusterAdminUser();
+    });
+
+    it('should show empty state when no GPU CQs exist or all are CPU-only', () => {
+      // Stage 1: no cluster queues at all
+      initIntercepts({ clusterQueues: [], cohortNames: [], resourceFlavors: [] });
+      infrastructurePage.visit();
+      infrastructurePage
+        .findCQUtilizationEmptyState()
+        .should('exist')
+        .should('contain.text', 'No accelerator cluster queues found');
+
+      // Stage 2: only CPU-only CQ — same empty state
+      initIntercepts({
+        clusterQueues: [
+          {
+            name: 'cq-cpu-only',
+            cohortName: 'cohort-1',
+            gpuFlavorName: undefined,
+            hasResourceGroups: true,
+          },
+        ],
+        cohortNames: ['cohort-1'],
+        resourceFlavors: [],
+      });
+      infrastructurePage.visit();
+      infrastructurePage.findCQUtilizationEmptyState().should('exist');
+    });
+
+    it('should render CQ card with subtitle, hardware badge, donut, workload counts, and cohort accordion', () => {
+      initIntercepts({
+        clusterQueues: [
+          {
+            name: 'cq-gpu',
+            cohortName: 'cohort-1',
+            gpuFlavorName: 'a100-flavor',
+            gpuNominalQuota: 8,
+            gpuUsed: 5,
+            admittedWorkloads: 2,
+            pendingWorkloads: 1,
+          },
+        ],
+        cohortNames: ['cohort-1'],
+        resourceFlavors: [{ name: 'a100-flavor', gpuProduct: 'NVIDIA A100' }],
+      });
+      infrastructurePage.visit();
+      infrastructurePage
+        .findCQUtilizationSubtitle()
+        .should('contain.text', 'Cluster queue accelerator utilization grouped by Kueue cohort.');
+      infrastructurePage.findCohortAccordion('cohort-1').should('exist');
+      infrastructurePage.findCQCard('cq-gpu').should('exist');
+      infrastructurePage.findHardwareModelBadge('NVIDIA A100').should('exist');
+      infrastructurePage.findAcceleratorDonutChart().should('exist');
+      infrastructurePage
+        .findCQWorkloadCounts()
+        .should('contain.text', '2 active workloads · 1 pending');
+    });
+
+    it('should render two CQ cards with DCGM utilization columns when telemetry is available', () => {
+      initIntercepts({
+        clusterQueues: [
+          {
+            name: 'notebook-queues',
+            cohortName: 'research-sandbox',
+            gpuFlavorName: 'mi300x-flavor',
+            gpuNominalQuota: 6,
+            gpuUsed: 2,
+            admittedWorkloads: 2,
+            pendingWorkloads: 2,
+          },
+          {
+            name: 'experiment-queues',
+            cohortName: 'research-sandbox',
+            gpuFlavorName: 'mi300x-flavor',
+            gpuNominalQuota: 4,
+            gpuUsed: 0,
+            admittedWorkloads: 0,
+            pendingWorkloads: 1,
+          },
+        ],
+        cohortNames: ['research-sandbox'],
+        resourceFlavors: [{ name: 'mi300x-flavor', gpuProduct: 'AMD MI300X' }],
+        dcgmModelName: 'AMD MI300X',
+      });
+      infrastructurePage.visit();
+      infrastructurePage.scrollToCQUtilizationSection();
+      infrastructurePage.findCohortAccordion('research-sandbox').should('exist');
+      infrastructurePage.findCQCard('notebook-queues').should('exist');
+      infrastructurePage.findCQCard('experiment-queues').should('exist');
+      // Scope to one card since both share the same model badge testid
+      infrastructurePage
+        .findCQCard('notebook-queues')
+        .findByTestId('hardware-model-badge-AMD MI300X')
+        .should('exist');
+      infrastructurePage.findAcceleratorDonutChartInCard('notebook-queues').should('exist');
+      infrastructurePage.findAcceleratorDonutChartInCard('experiment-queues').should('exist');
+    });
+
+    describe('borrow-lend donut state', () => {
+      // a100-train-queues: 6 nominal, 4 used → lends 2; burst-training: 8 nominal, 10 used, 2 borrowed
+      const COHORT = 'ml-training-cohort';
+      const FLAVOR = 'a100-flavor';
+
+      beforeEach(() => {
+        initIntercepts({
+          clusterQueues: [
+            {
+              name: 'a100-train-queues',
+              cohortName: COHORT,
+              gpuFlavorName: FLAVOR,
+              gpuNominalQuota: 6,
+              gpuUsed: 4,
+              admittedWorkloads: 1,
+              pendingWorkloads: 2,
+            },
+            {
+              name: 'burst-training',
+              cohortName: COHORT,
+              gpuFlavorName: FLAVOR,
+              gpuNominalQuota: 8,
+              gpuUsed: 10,
+              gpuBorrowed: 2,
+              admittedWorkloads: 2,
+              pendingWorkloads: 2,
+            },
+          ],
+          cohortNames: [COHORT],
+          resourceFlavors: [{ name: FLAVOR, gpuProduct: 'NVIDIA A100' }],
+          dcgmModelName: 'NVIDIA A100',
+        });
+        infrastructurePage.visit();
+        infrastructurePage.scrollToCQUtilizationSection();
+      });
+
+      it('shows lend/borrow badges, workload counts, all chart columns, and per-model badge popovers', () => {
+        // Cohort-level badges
+        infrastructurePage.findCohortAccordion(COHORT).should('exist');
+        infrastructurePage.findCohortBorrowLendBadge().should('exist');
+        infrastructurePage
+          .findCohortUnallocatedBorrowable()
+          .should('contain.text', '2 unallocated, borrowable');
+
+        // Lender card
+        infrastructurePage
+          .findCQLendBadgeInCard('a100-train-queues')
+          .should('contain.text', '2 lent');
+        infrastructurePage
+          .findWorkloadCountsInCard('a100-train-queues')
+          .should('contain.text', '1 active workload · 2 pending');
+        infrastructurePage.findAcceleratorDonutChartInCard('a100-train-queues').should('exist');
+        infrastructurePage
+          .findCQCard('a100-train-queues')
+          .should('contain.text', 'Accelerator compute utilization')
+          .should('contain.text', 'Accelerator memory utilization');
+
+        // Lent badge popover: shows counterpart CQ and per-model lent count
+        infrastructurePage.findCQLendBadgeInCard('a100-train-queues').click();
+        infrastructurePage
+          .findOpenPopover()
+          .should('contain.text', 'Lent capacity')
+          .should('contain.text', 'burst-training')
+          .should('contain.text', '2 × NVIDIA A100');
+        cy.get('body').type('{esc}');
+
+        // Borrower card
+        infrastructurePage
+          .findCQBorrowBadgeInCard('burst-training')
+          .should('contain.text', '+2 borrowed');
+        infrastructurePage
+          .findWorkloadCountsInCard('burst-training')
+          .should('contain.text', '2 active workloads · 2 pending');
+        infrastructurePage.findAcceleratorDonutChartInCard('burst-training').should('exist');
+        infrastructurePage
+          .findCQCard('burst-training')
+          .should('contain.text', 'Accelerator compute utilization')
+          .should('contain.text', 'Accelerator memory utilization');
+
+        // Borrowed badge popover: shows counterpart CQ and per-model borrowed count
+        infrastructurePage.findCQBorrowBadgeInCard('burst-training').click();
+        infrastructurePage
+          .findOpenPopover()
+          .should('contain.text', 'Borrowed capacity')
+          .should('contain.text', 'a100-train-queues')
+          .should('contain.text', '2 × NVIDIA A100');
+        cy.get('body').type('{esc}');
+      });
+
+      it('shows per-model breakdown in donut segment hover tooltip', () => {
+        // Total accelerators — Own segment
+        infrastructurePage
+          .findAcceleratorDonutChartInCard('a100-train-queues')
+          .find('svg path')
+          .first()
+          .trigger('mouseover', { force: true });
+        infrastructurePage
+          .findCQCard('a100-train-queues')
+          .should('contain.text', 'NVIDIA A100: 4/6 in use');
+
+        // Total accelerators — Lent segment
+        infrastructurePage
+          .findAcceleratorDonutChartInCard('a100-train-queues')
+          .find('svg path')
+          .eq(1)
+          .trigger('mouseover', { force: true });
+        infrastructurePage
+          .findCQCard('a100-train-queues')
+          .should('contain.text', '2 GPUs lent')
+          .should('contain.text', 'NVIDIA A100: 2 lent');
+
+        // DCGM compute — Lent segment
+        infrastructurePage
+          .findDcgmComputeDonutInCard('a100-train-queues')
+          .find('svg path')
+          .eq(1)
+          .trigger('mouseover', { force: true });
+        infrastructurePage
+          .findCQCard('a100-train-queues')
+          .should('contain.text', 'NVIDIA A100: 30% utilization');
+      });
+
+      describe('pure borrower CQ (nominal=0, used>0)', () => {
+        const PURE_BORROWER_COHORT = 'burst-cohort';
+        const PURE_BORROWER_FLAVOR = 'h100-flavor';
+
+        beforeEach(() => {
+          initIntercepts({
+            clusterQueues: [
+              {
+                name: 'pure-borrower-cq',
+                cohortName: PURE_BORROWER_COHORT,
+                gpuFlavorName: PURE_BORROWER_FLAVOR,
+                gpuNominalQuota: 0,
+                gpuUsed: 6,
+                admittedWorkloads: 1,
+                pendingWorkloads: 0,
+              },
+            ],
+            cohortNames: [PURE_BORROWER_COHORT],
+            resourceFlavors: [{ name: PURE_BORROWER_FLAVOR, gpuProduct: 'NVIDIA H100' }],
+            dcgmModelName: 'NVIDIA H100',
+          });
+          infrastructurePage.visit();
+        });
+
+        it('renders the CQ card with all 3 chart columns and a fully-filled borrowed donut', () => {
+          infrastructurePage.findAcceleratorDonutChartInCard('pure-borrower-cq').should('exist');
+          infrastructurePage
+            .findCQCard('pure-borrower-cq')
+            .findByText('+6 borrowed')
+            .should('exist');
+          infrastructurePage
+            .findCQCard('pure-borrower-cq')
+            .should('contain.text', 'Accelerator compute utilization')
+            .should('contain.text', 'Accelerator memory utilization');
+        });
+      });
+    });
+  });
 });
